Update template model

+++ document_merge_service/api/models.py
@@ -29,6 +29,24 @@ class Template(models.Model):
     engine: models.CharField = models.CharField(
         max_length=20, choices=ENGINE_CHOICES_TUPLE
     )
+    created_at: models.DateTimeField = models.DateTimeField(
+        auto_now_add=True, db_index=True, null=True
+    )
+    created_by_user: models.CharField = models.CharField(
+        max_length=150, blank=True, null=True
+    )
+    created_by_group: models.CharField = models.CharField(
+        max_length=255, blank=True, null=True
+    )
+    modified_at: models.DateTimeField = models.DateTimeField(
+        auto_now=True, db_index=True, null=True
+    )
+    modified_by_user: models.CharField = models.CharField(
+        max_length=150, blank=True, null=True
+    )
+    modified_by_group: models.CharField = models.CharField(
+        max_length=255, blank=True, null=True
+    )
     meta = models.JSONField(default=dict)
 
 
@@ -36,7 +54,8 @@ class Template(models.Model):
 def auto_delete_file_on_delete(sender, instance, **kwargs):
     """Delete template file from filesystem when `Template` object is deleted."""
 
-    DefaultStorage().delete(instance.template.name)
+    if instance.template:
+        DefaultStorage().delete(instance.template.name)
 
 
 @receiver(models.signals.pre_save, sender=Template)
